Move functions to new utils file in src/
